refactor: consolidate trivial auto-claude steps into simple-steps.ts
Merge near-identical plan/simplify/review wrappers into one file.
Keep implement.ts and create-pr.ts which have real logic.
