+++ src/lib/auto-claude/pipeline.ts
@@ -6,9 +6,7 @@ import { ARTIFACTS } from "./prompt-templates/index.js";
 import type { StepName } from "./prompt-templates/index.js";
 import { createPr } from "./steps/create-pr.js";
 import { stepImplement } from "./steps/implement.js";
-import { stepPlan } from "./steps/plan.js";
-import { stepReview } from "./steps/review.js";
-import { stepSimplify } from "./steps/simplify.js";
+import { stepPlan, stepReview, stepSimplify } from "./steps/simple-steps.js";
 import { LABELS, ensureLabelsExist, removeLabel, setLabel } from "./labels.js";
 import { ensureDir, fileExists, readFile, writeFile } from "../../utils/fs.js";
 import { execSafe, git } from "../../utils/git/exec.js";
